basictools/backend/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
python-dotenv==0.21.0
pillow==9.5.0
aiofiles==0.8.0
